修改格式刷属性
